Bcc.CLI.Cole.Delegation

Documentation

data ColeDelegationError

Constructors

CertificateValidationErrors !FilePath ![Text]

DlgCertificateDeserialisationFailed !FilePath !Text

ColeDelegationKeyError !ColeKeyFailure

checkColeGenesisDelegation ∷ CertificateFile → ProtocolMagicId → VerificationKey → VerificationKey → ExceptT ColeDelegationError IO ()

Verify that a certificate signifies genesis delegation by assumed genesis key to a delegate key, for a given protocol magic. If certificate fails validation, throw an error.

issueColeGenesisDelegation ∷ ProtocolMagicId → EpochNumber → SigningKey → VerificationKey → Certificate

Issue a certificate for genesis delegation to a delegate key, signed by the issuer key, for a given protocol magic and coming into effect at given epoch.

renderColeDelegationError ∷ ColeDelegationError → Text

serialiseDelegationCert ∷ Certificate → ByteString

serialiseColeWitness ∷ SomeColeSigningKey → ByteString