Bcc.CLI.Cole.Tx

Documentation

data ColeTxError

Constructors

TxDeserialisationFailed !FilePath !DecoderError

ColeTxSubmitError !Text

ColeTxSubmitErrorEraMismatch !EraMismatch

EnvSocketError !EnvSocketError

newtype TxFile

Constructors

TxFile FilePath

newtype NewTxFile

Constructors

NewTxFile FilePath

prettyAddress ∷ Address ColeAddr → Text

Pretty-print an address in its Base58 form, and also its full structure.

readColeTx ∷ TxFile → ExceptT ColeTxError IO (ATxAux ByteString)

normalColeTxToGenTx ∷ ATxAux ByteString → GenTx ColeBlock

The GenTx is all the kinds of transactions that can be submitted and "normal" Cole transactions are just one of the kinds.

txSpendGenesisUTxOColePBFT ∷ Config → NetworkId → SomeColeSigningKey → Address ColeAddr → [TxOut ColeEra] → Tx ColeEra

Generate a transaction spending genesis UTxO at a given address, to given outputs, signed by the given key.

txSpendUTxOColePBFT ∷ NetworkId → SomeColeSigningKey → [TxIn] → [TxOut ColeEra] → Tx ColeEra

Generate a transaction from given Tx inputs to outputs, signed by the given key.

nodeSubmitTx ∷ NetworkId → GenTx ColeBlock → ExceptT ColeTxError IO ()

Submit a transaction to a node specified by topology info.

renderColeTxError ∷ ColeTxError → Text

fromCborTxAux ∷ ByteString → Either DecoderError (ATxAux ByteString)

toCborTxAux ∷ ATxAux ByteString → ByteString

data ScriptValidity

Indicates whether a script is expected to fail or pass validation.

Constructors

ScriptInvalid
Script is expected to fail validation. Transactions marked as such can include scripts that fail validation. Such transactions may be submitted to the chain, in which case the collateral will be taken upon on chain script validation failure.

ScriptValid
Script is expected to pass validation. Transactions marked as such cannot include scripts that fail validation.