Bcc.CLI.Sophie.Commands

Description

Sophie CLI command types

CLI command types

data SophieCommand

All the CLI subcommands under "sophie".

Constructors

AddressCmd AddressCmd

StakeAddressCmd StakeAddressCmd

KeyCmd KeyCmd

TransactionCmd TransactionCmd

NodeCmd NodeCmd

PoolCmd PoolCmd

QueryCmd QueryCmd

GovernanceCmd GovernanceCmd

GenesisCmd GenesisCmd

TextViewCmd TextViewCmd

data AddressCmd

Constructors

AddressKeyGen AddressKeyType VerificationKeyFile SigningKeyFile

AddressKeyHash VerificationKeyTextOrFile (Maybe OutputFile)

AddressBuild PaymentVerifier (Maybe StakeVerifier) NetworkId (Maybe OutputFile)

AddressBuildMultiSig ScriptFile NetworkId (Maybe OutputFile)

AddressInfo Text (Maybe OutputFile)

data StakeAddressCmd

Constructors

StakeAddressKeyGen VerificationKeyFile SigningKeyFile

StakeAddressKeyHash (VerificationKeyOrFile StakeKey) (Maybe OutputFile)

StakeAddressBuild (VerificationKeyOrFile StakeKey) NetworkId (Maybe OutputFile)

StakeRegistrationCert StakeVerifier OutputFile

StakeCredentialDelegationCert StakeVerifier (VerificationKeyOrHashOrFile StakePoolKey) OutputFile

StakeCredentialDeRegistrationCert StakeVerifier OutputFile

data KeyCmd

Constructors

KeyGetVerificationKey SigningKeyFile VerificationKeyFile

KeyNonExtendedKey VerificationKeyFile VerificationKeyFile

KeyConvertColeKey (Maybe Text) ColeKeyType SomeKeyFile OutputFile

KeyConvertColeGenesisVKey VerificationKeyBase64 OutputFile

KeyConvertColeVestedVKey VerificationKeyBase64 OutputFile

KeyConvertITNStakeKey SomeKeyFile OutputFile

KeyConvertITNExtendedToStakeKey SomeKeyFile OutputFile

KeyConvertITNBip32ToStakeKey SomeKeyFile OutputFile

KeyConvertBccAddressSigningKey BccAddressKeyType SigningKeyFile OutputFile

data TransactionCmd

Constructors

TxBuildRaw

Fields

AnyBccEra

(Maybe ScriptValidity)

Mark script as expected to pass or fail validation

[(TxIn, Maybe (ScriptWitnessFiles WitCtxTxIn))]

Transaction inputs with optional spending scripts

[TxIn]

Transaction inputs for collateral, only key witnesses, no scripts.

[WitnessSigningData]

Required signers

[TxOutAnyEra]

(Maybe (Value, [ScriptWitnessFiles WitCtxMint]))

Multi-Asset value with script witness

(Maybe SlotNo)

Transaction lower bound

(Maybe SlotNo)

Transaction upper bound

(Maybe Entropic)

Tx fee

[(CertificateFile, Maybe (ScriptWitnessFiles WitCtxStake))]

Certificates with potential script witness

[(StakeAddress, Entropic, Maybe (ScriptWitnessFiles WitCtxStake))]

TxMetadataJsonSchema

[ScriptFile]

Auxillary scripts

[MetadataFile]

(Maybe ProtocolParamsSourceSpec)

(Maybe UpdateProposalFile)

TxBodyFile

TxBuild
Like TxBuildRaw but without the fee, and with a change output.

Fields

AnyBccEra

AnyConsensusModeParams

NetworkId

(Maybe ScriptValidity)

Mark script as expected to pass or fail validation

(Maybe Word)

Override the required number of tx witnesses

[(TxIn, Maybe (ScriptWitnessFiles WitCtxTxIn))]

Required signers

[WitnessSigningData]

Transaction inputs with optional spending scripts

[TxIn]

Transaction inputs for collateral, only key witnesses, no scripts.

[TxOutAnyEra]

Normal outputs

TxOutChangeAddress

A change output

(Maybe (Value, [ScriptWitnessFiles WitCtxMint]))

Multi-Asset value with script witness

(Maybe SlotNo)

Transaction lower bound

(Maybe SlotNo)

Transaction upper bound

[(CertificateFile, Maybe (ScriptWitnessFiles WitCtxStake))]

Certificates with potential script witness

[(StakeAddress, Entropic, Maybe (ScriptWitnessFiles WitCtxStake))]

Withdrawals with potential script witness

TxMetadataJsonSchema

[ScriptFile]

Auxillary scripts

[MetadataFile]

(Maybe ProtocolParamsSourceSpec)

(Maybe UpdateProposalFile)

TxBodyFile

TxSign TxBodyFile [WitnessSigningData] (Maybe NetworkId) TxFile

TxCreateWitness TxBodyFile WitnessSigningData (Maybe NetworkId) OutputFile

TxAssembleTxBodyWitness TxBodyFile [WitnessFile] OutputFile

TxSubmit AnyConsensusModeParams NetworkId FilePath

TxMintedPolicyId ScriptFile

TxCalculateMinFee TxBodyFile (Maybe NetworkId) ProtocolParamsSourceSpec TxInCount TxOutCount TxSophieWitnessCount TxColeWitnessCount

TxCalculateMinRequiredUTxO AnyBccEra ProtocolParamsSourceSpec TxOutAnyEra

TxHashScriptData ScriptDataOrFile

TxGetTxId InputTxFile

TxView InputTxFile

data NodeCmd

Constructors

NodeKeyGenCold VerificationKeyFile SigningKeyFile OpCertCounterFile

NodeKeyGenKES VerificationKeyFile SigningKeyFile

NodeKeyGenVRF VerificationKeyFile SigningKeyFile

NodeKeyHashVRF (VerificationKeyOrFile VrfKey) (Maybe OutputFile)

NodeNewCounter ColdVerificationKeyOrFile Word OpCertCounterFile

NodeIssueOpCert (VerificationKeyOrFile KesKey) SigningKeyFile OpCertCounterFile KESPeriod OutputFile

data PoolCmd

Constructors

PoolRegistrationCert

Fields

(VerificationKeyOrFile StakePoolKey)

Stake pool verification key.

(VerificationKeyOrFile VrfKey)

VRF Verification key.

Entropic

Pool pledge.

Entropic

Pool cost.

Rational

Pool margin.

(VerificationKeyOrFile StakeKey)

Reward account verification staking key.

[VerificationKeyOrFile StakeKey]

Pool owner verification staking key(s).

[StakePoolRelay]

Stake pool relays.

(Maybe StakePoolMetadataReference)

Stake pool metadata.

NetworkId

OutputFile

PoolRetirementCert

Fields

(VerificationKeyOrFile StakePoolKey)

Stake pool verification key.

EpochNo

Epoch in which to retire the stake pool.

OutputFile

PoolGetId (VerificationKeyOrFile StakePoolKey) OutputFormat

PoolMetadataHash PoolMetadataFile (Maybe OutputFile)

data QueryCmd

Constructors

QueryProtocolParameters' AnyConsensusModeParams NetworkId (Maybe OutputFile)

QueryTip AnyConsensusModeParams NetworkId (Maybe OutputFile)

QueryStakePools' AnyConsensusModeParams NetworkId (Maybe OutputFile)

QueryStakeDistribution' AnyConsensusModeParams NetworkId (Maybe OutputFile)

QueryStakeAddressInfo AnyConsensusModeParams StakeAddress NetworkId (Maybe OutputFile)

QueryUTxO' AnyConsensusModeParams QueryUTxOFilter NetworkId (Maybe OutputFile)

QueryDebugLedgerState' AnyConsensusModeParams NetworkId (Maybe OutputFile)

QueryProtocolState' AnyConsensusModeParams NetworkId (Maybe OutputFile)

QueryStakeSnapshot' AnyConsensusModeParams NetworkId (Hash StakePoolKey)

QueryPoolParams' AnyConsensusModeParams NetworkId (Hash StakePoolKey)

data GovernanceCmd

Constructors

GovernanceMIRPayStakeAddressesCertificate MIRPot [StakeAddress] [Entropic] OutputFile

GovernanceMIRTransfer Entropic OutputFile TransferDirection

GovernanceGenesisKeyDelegationCertificate (VerificationKeyOrHashOrFile GenesisKey) (VerificationKeyOrHashOrFile GenesisDelegateKey) (VerificationKeyOrHashOrFile VrfKey) OutputFile

GovernanceVestedKeyDelegationCertificate (VerificationKeyOrHashOrFile VestedKey) (VerificationKeyOrHashOrFile VestedDelegateKey) (VerificationKeyOrHashOrFile VrfKey) OutputFile

GovernanceUpdateProposal OutputFile EpochNo [VerificationKeyFile] ProtocolParametersUpdate

data GenesisCmd

Constructors

GenesisCreate GenesisDir Word Word Word (Maybe SystemStart) (Maybe Entropic) NetworkId

GenesisCreateStaked GenesisDir Word Word Word Word Word (Maybe SystemStart) (Maybe Entropic) Entropic NetworkId Word Word Word

GenesisKeyGenGenesis VerificationKeyFile SigningKeyFile

GenesisKeyGenDelegate VerificationKeyFile SigningKeyFile OpCertCounterFile

GenesisKeyGenVested VerificationKeyFile SigningKeyFile

GenesisKeyGenVestedDelegate VerificationKeyFile SigningKeyFile OpCertCounterFile

GenesisKeyGenUTxO VerificationKeyFile SigningKeyFile

GenesisCmdKeyHash VerificationKeyFile

GenesisVerKey VerificationKeyFile SigningKeyFile

GenesisTxIn VerificationKeyFile NetworkId (Maybe OutputFile)

GenesisAddr VerificationKeyFile NetworkId (Maybe OutputFile)

GenesisHashFile GenesisFile

data TextViewCmd

Constructors

TextViewInfo !FilePath (Maybe OutputFile)

renderSophieCommand ∷ SophieCommand → Text

CLI flag types

data AddressKeyType

Constructors

AddressKeySophie

AddressKeySophieExtended

AddressKeyCole

data ColeKeyType

Constructors

ColePaymentKey ColeKeyFormat

ColeGenesisKey ColeKeyFormat

ColeGenesisVestedKey ColeKeyFormat

ColeVestedKey ColeKeyFormat

ColeDelegateKey ColeKeyFormat

ColeVestedDelegateKey ColeKeyFormat

data ColeKeyFormat

Constructors

NonLegacyColeKeyFormat

LegacyColeKeyFormat

data BccAddressKeyType

The type of bcc-address key.

Constructors

BccAddressSophiePaymentKey

BccAddressSophieStakeKey

BccAddressIcarusPaymentKey

BccAddressColePaymentKey

newtype GenesisDir

Constructors

GenesisDir FilePath

newtype VestedDir

Constructors

VestedDir FilePath

newtype TxInCount

Constructors

TxInCount Int

newtype TxOutCount

Constructors

TxOutCount Int

newtype TxSophieWitnessCount

Constructors

TxSophieWitnessCount Int

newtype TxColeWitnessCount

Constructors

TxColeWitnessCount Int

data SomeKeyFile

Either a verification or signing key, used for conversions and other commands that make sense for both.

Constructors

AVerificationKeyFile VerificationKeyFile

ASigningKeyFile SigningKeyFile

newtype OpCertCounterFile

Constructors

OpCertCounterFile FilePath

newtype OutputFile

Constructors

OutputFile FilePath

newtype ProtocolParamsFile

Constructors

ProtocolParamsFile FilePath

data ProtocolParamsSourceSpec

Constructors

ParamsFromGenesis !GenesisFile
We allow an appropriately forewarned user to obtain protocol params directly from the genesis file, which allows them to avoid running the node in case they would like to estimate the fee using the blockchain's initial protocol parameters.

ParamsFromFile !ProtocolParamsFile
Obtain protocol parameters from a file structured by the 'bcc-api' ProtocolParameters data type.

newtype WitnessFile

Constructors

WitnessFile FilePath

newtype TxBodyFile

Constructors

TxBodyFile FilePath

newtype TxFile

Constructors

TxFile FilePath

data InputTxFile

Constructors

InputTxBodyFile TxBodyFile

InputTxFile TxFile

newtype VerificationKeyBase64

A raw verification key given in Base64, and decoded into a ByteString.

Constructors

VerificationKeyBase64 String

newtype GenesisKeyFile

Constructors

GenesisKeyFile FilePath

newtype VestedKeyFile

Constructors

VestedKeyFile FilePath

data MetadataFile

Constructors

MetadataFileJSON FilePath

MetadataFileCBOR FilePath

newtype PoolId

Constructors

PoolId String

newtype PoolMetadataFile

Constructors

PoolMetadataFile

Fields

unPoolMetadataFile ∷ FilePath

newtype PrivKeyFile

Constructors

PrivKeyFile FilePath

newtype BlockId

Constructors

BlockId String

data WitnessSigningData

Data required to construct a witness.

Constructors

KeyWitnessSigningData

Fields

!SigningKeyFile

Path to a file that should contain a signing key.

!(Maybe (Address ColeAddr))

An optionally specified Cole address.

If specified, both the network ID and derivation path are extracted from the address and used in the construction of the Cole witness.

data ColdVerificationKeyOrFile

Either a stake pool verification key, genesis delegate verification key, or a path to a cold verification key file.

Note that a "cold verification key" refers to either a stake pool or genesis delegate verification key.

TODO: A genesis delegate extended key should also be valid here.

Constructors

ColdStakePoolVerificationKey !(VerificationKey StakePoolKey)

ColdGenesisDelegateVerificationKey !(VerificationKey GenesisDelegateKey)

ColdGenesisVestedDelegateVerificationKey !(VerificationKey GenesisVestedDelegateKey)

ColdVestedDelegateVerificationKey !(VerificationKey VestedDelegateKey)

ColdVerificationKeyFile !VerificationKeyFile